+++ ffi/capi/src/lib.rs
@@ -82,6 +82,8 @@ pub mod collections_sets;
 pub mod date;
 #[cfg(feature = "datetime")]
 pub mod date_formatter;
+#[cfg(feature = "datetime")]
+pub mod date_time_formatter;
 #[cfg(any(feature = "datetime", feature = "timezone", feature = "calendar"))]
 pub mod datetime;
 #[cfg(feature = "datetime")]
@@ -104,8 +106,6 @@ pub mod list;
 pub mod locale;
 #[cfg(feature = "locale")]
 pub mod locale_directionality;
-#[cfg(feature = "datetime")]
-pub mod neo_datetime;
 #[cfg(feature = "normalizer")]
 pub mod normalizer;
 #[cfg(feature = "normalizer")]
@@ -154,6 +154,8 @@ pub mod week;
 pub mod windows_parser;
 #[cfg(feature = "datetime")]
 pub mod zoned_date_formatter;
+#[cfg(feature = "datetime")]
+pub mod zoned_date_time_formatter;
 #[cfg(feature = "datetime")]
 pub mod zoned_datetime;
 #[cfg(feature = "datetime")]
